add utils.go in x509 directory
